Refactor part of RuntimeFilter's code
