+++ be/src/exec/hash_join_node.h
@@ -25,6 +25,7 @@
 
 #include "exec/exec_node.h"
 #include "exec/hash_table.h"
+#include "exprs/runtime_filter_slots.h"
 #include "gen_cpp/PlanNodes_types.h"
 
 namespace doris {
